requirements.txt
dnspython==2.3.0
psycopg2-binary==2.9.6
pymongo==4.3.3